// ai-assistant/projects/ai-assistant/src/lib/enums/numbers.enum.ts
export enum Integers {
  Zero = 0,
  One,
  Two,
  Three,
  Four,
  Five,
  Six,
  Seven,
  Eight,
  Nine,
  Ten = 10,
  Eleven,
  Twelve,
  Thirteen,
  Fourteen = 14,
  Fifteen = 15,
  Seventeen = 17,
  Eighteen = 18,
  Nineteen = 19,
  Twenty = 20,
  TwentyOne = 21,
  TwentyThree = 23,
  TwentyFour = 24,
  TwentyFive,
  TwentySix,
  TwentySeven,
  TwentyEight = 28,
  TwentyNine = 29,
  Thirty,
  ThirtyOne,
  ThirtyTwo,
  ThirtyThree,
  ThirtyFive = 35,
  ThirtySix = 36,
  Forty = 40,
  FortyTwo = 42,
  FortyFour = 44,
  FortyFive = 45,
  FortySix,
  FortyEight = 48,
  Fifty = 50,
  FiftyTwo = 52,
  FiftySix = 56,
  FiftySeven = 57,
  FiftyNine = 59,
  Sixty,
  SixtyThree = 63,
  SixtyFour = 64,
  SixtyFive = 65,
  SixtySeven = 67,
  Seventy = 70,
  SeventyThree = 73,
  SeventySix = 76,
  Eighty = 80,
  EightyFour = 84,
  EightySix = 86,
  NinetyFive = 95,
  OneHundred = 100,
  OneHundredThree = 103,
  OneHundredTen = 110,
  OneHundredFifteen = 115,
  OneHundredSixteen = 116,
  OneHundredTwenty = 120,
  OneHundredForty = 140,
  OneHundredFifty = 150,
  OneHundredSixty = 160,
  OneHundredSixtyTwo = 162,
  OneHundredSeventy = 170,
  OneHundredEightyFive = 185,
  OneHundredNinety = 190,
  TwoHundred = 200,
  TwoHundredFifteen = 215,
  TwoHundredTwentyFive = 225,
  TwoHundredFifty = 250,
  TwoHundredSixty = 260,
  TwoHundredEighty = 280,
  ThreeHundred = 300,
  ThreeHundredFifty = 350,
  ThreeHundredSixtyFive = 365,
  FourHundred = 400,
  FourHundredFour = 404,
  FourHundredTwenty = 420,
  FourHundredThirty = 430,
  FourHundredFourtyEight = 448,
  FourHundredFourtySix = 446,
  FiveHundred = 500,
  SevenHundredThirty = 730,
  OneThousand = 1000,
  OneThousandThreeHundredEighty = 1380,
  OneThousandEightHundred = 1800,
  TwoThousandNineteen = 2019,
  TwoThousandTwenty = 2020,
  TwoThousandTwentyOne = 2021,
  TwoThousandTwentyTwo = 2022,
  TwoThousandTwentyThree = 2023,
  TwoThousandTwentyFour = 2024,
  ThreeThousandAndSixHundred = 3600,
  TenThousand = 10000,
  EightSixThousandFourHundred = 86400,
  Ninety = 90,
  OneMillion = 1000000,
  OneBillion = 1000000000,
  OneTrillion = 1000000000000,
}